Hexagon: add hardware prefetch settings
dcache prefetch is in a user-mode register which might not always be set
for best performance.

When switching into the kernel, the kernel always sets it for itself to
improve performance.  This gets restored on the way back out.

For processes, a Kconfig switch is added to initialize the register if
desired.

Signed-off-by: Richard Kuo <rkuo@codeaurora.org>