Hexagon: add v5 hardware timer support
The hardware timer changed somewhat with later architectures;
it's 64bit, we don't really need to do any more of the
clear/enable stuff, etc.  Also cleaned up a few things left over
from the old driver like platform cruft and mixed use of
pcycles/hypervisor calls.

Although normally we want to use the hypervisor's interface,
this is available if the hypervisor allows it.

Signed-off-by: Richard Kuo <rkuo@codeaurora.org>